requirements.txt

telethon>=1.28.0


httpx>=0.24.0


PyYAML>=6.0


aiofiles>=23.0.0


loguru>=0.7.0


psutil>=5.9.0


python-dateutil>=2.8.0
pytz>=2023.3


unidecode>=1.3.0


python-dotenv>=1.0.0


pydantic>=2.4.0


beautifulsoup4>=4.12.0


requests>=2.28.0


cryptography>=3.4.8


tenacity>=8.2.0


schedule>=1.2.0


anyio>=3.0.0
certifi
h11>=0.14.0
sniffio


aiosqlite>=0.19.0


requests>=2.31.0
httpx>=0.24.0
